fix: hashpack install re-rendering issue (#244)
till now we didn't have a way to re-render wallet modal if we detect hashpack is installed after someone open the modal. to handle it we need to re-render wallet modal as soon as we know hashpack is available. so to solve that issue I have installed event-emitter package which will emit event from HashConnector class to WalletModal component and WalletModal will listen to that event, so we can set some boolean in useState to toggle install state.